Clean up import methods into separate file
