+++ src/armature/Model.ts
@@ -1,13 +1,12 @@
 import { mat3, mat4, vec3, vec4 } from 'gl-matrix';
 
 import { Color } from '../colors/Color';
-import { RGBColor } from '../colors/RGBColor';
 import { AABB } from '../geometry/BakedGeometry';
-import { Face, WorkingGeometry } from '../geometry/WorkingGeometry';
 import { MtlData, ObjData } from '../import_dir/ImportData';
-import { BakedMaterial, Material } from '../renderer/Material';
+import { BakedMaterial } from '../renderer/Material';
 import { RenderObject } from '../types/RenderObject';
 import { worldSpaceAABB } from '../utils/aabb';
+import { importOBJ } from './Import';
 import { GeometryNode, Node } from './Node';
 import { NodeRenderObject } from './NodeRenderObject';
 
@@ -58,129 +57,9 @@ export class Model {
      * import_dir directory.
      */
     public static importOBJ() {
-        let lines = ObjData.split('\n');
-
-        // Hack
-        lines[0].split(' '); // [1] is the file name
-        lines = lines.slice(1);
-        lines[0].split(' '); // [1] is the mtlFileName
-        lines = lines.slice(1);
-
-        const vertices: vec3[] = [];
-        while (lines[0].startsWith('v ')) {
-            const vecLine = lines[0]
-                .split(' ')
-                .slice(1)
-                .map(parseFloat);
-            const vec = vec3.fromValues(vecLine[0], vecLine[1], vecLine[2]);
-            vertices.push(vec);
-            lines.shift();
-        }
-
-        const normals: vec3[] = [];
-        while (lines[0].startsWith('vn')) {
-            const vecLine = lines[0]
-                .split(' ')
-                .slice(1)
-                .map(parseFloat);
-            const vec = vec3.fromValues(vecLine[0], vecLine[1], vecLine[2]);
-            normals.push(vec);
-            lines.shift();
-        }
-
-        /**
-         * A helper class that specifies a group as stored in the obj files
-         */
-        class Group {
-            public groupName: string;
-            public materialName: string;
-            public faces: Face[];
-        }
-
-        const groups: Group[] = [];
-        while (lines.length > 0 && lines[0].startsWith('g')) {
-            const groupName = lines[0].split(' ')[1];
-            lines = lines.slice(1);
-            const materialName = lines[0].split(' ')[1];
-            lines = lines.slice(1);
-            const faces: Face[] = [];
-            while (lines.length > 0 && lines[0].startsWith('f')) {
-                const index: string[][] = lines[0]
-                    .split(' ')
-                    .slice(1)
-                    .map((s: string) => s.split('//'));
-                // FIXME(pbardea): How to concat three arrays without needing empty array hack.
-                const empty: string[] = [];
-                const face: number[] = empty
-                    .concat(...index)
-                    .filter((_: string, idx: number) => idx % 2 === 0)
-                    .map((s: string) => parseInt(s, 10) - 1);
-                faces.push(new Face(face));
-                lines = lines.slice(1);
-            }
-            groups.push({
-                groupName: groupName,
-                materialName: materialName,
-                faces: faces
-            });
-        }
-
-        lines = MtlData.split('\n');
-        const materials: Map<string, Material> = new Map<string, Material>();
-        while (lines.length > 0 && lines[0].startsWith('newmtl')) {
-            const materialName = lines[0].split(' ')[1];
-            lines = lines.slice(1);
-            // Assume the prefixes of the following lines are Ka, Kd, Ks, Ns, illum
-            lines[0]
-                .split(' ')
-                .slice(1)
-                .map(parseFloat); // ka
-            lines = lines.slice(1);
-            const kd = lines[0]
-                .split(' ')
-                .slice(1)
-                .map(parseFloat);
-            lines = lines.slice(1);
-            lines[0]
-                .split(' ')
-                .slice(1)
-                .map(parseFloat); // ks
-            lines = lines.slice(1);
-            const ns = parseFloat(lines[0].split(' ')[1]);
-            lines = lines.slice(1);
-            // Ignore the illumination for now
-            lines = lines.slice(1);
-
-            materials.set(
-                materialName,
-                Material.create({
-                    color: RGBColor.fromRGB(kd[0] * 255, kd[1] * 255, kd[2] * 255),
-                    shininess: ns
-                })
-            );
-        }
-
-        const parent = new Node();
-        parent.setAnchor(vec3.fromValues(0, 0, 0));
-
-        const geoNodes: GeometryNode[] = groups.map(
-            (group: Group) =>
-                new GeometryNode(
-                    new WorkingGeometry({
-                        vertices: vertices,
-                        normals: normals,
-                        faces: group.faces,
-                        material: <Material>materials.get(group.materialName),
-                        controlPoints: []
-                    }),
-                    parent
-                )
-        );
-        geoNodes.forEach((node: Node) => node.setAnchor(vec3.fromValues(0, 0, 0)));
-
-        return new Model([parent, ...geoNodes]);
+        const nodes = importOBJ(ObjData, MtlData);
+        return new Model(nodes);
     }
-
 
     /**
      * Creates a new model.
